requirements.txt
PyYAML>=6.0,<7.0
requests>=2.27.1,<3.0
